# Reduce allocations by avoiding strings.Split